requirements.txt
asgiref==3.5.2
Django==4.0.6
django-ckeditor==6.4.2
django-js-asset==2.0.0
django-taggit==3.0.0
Markdown==3.4.1
psycopg2-binary==2.9.3
sqlparse==0.4.2
